+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 ### Features
 
 - Postpone Update Api call integration ([PR 485](https://github.com/input-output-hk/daedalus/pull/485))
+- Wallet restore recovery phrase textarea replaced with React-Polymorph Autocomplete ([PR 516](https://github.com/input-output-hk/daedalus/pull/516))
 
 ### Fixes